Teach chanhistory to not repeat itself
This makes it so chanhistory (with the help of whowas) does not replay
messages that happened while you were connected to the server.

This is perhaps a 60% solution to this problem; it would probably be
better if it actually tracked when you were in the channel, not when you
were active on the server and then quit.